svg/types/scripted/SVGPoint.html

<!DOCTYPE HTML>
<title>SVGPoint interface</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  let svgElement = document.createElementNS("http://www.w3.org/2000/svg", "svg");
  let point = svgElement.createSVGPoint();

  // Check initial point values.
  assert_equals(point.x, 0);
  assert_equals(point.y, 0);

  // Check setting valid point values.
  point.x = 100;
  assert_equals(point.x, 100);
  point.y = 200;
  assert_equals(point.y, 200);

  // Check setting invalid arguments.
  assert_throws_js(TypeError, () => { point.x = point; });
  assert_throws_js(TypeError, () => { point.y = NaN; });
  point.y = null;
  assert_equals(point.y, 0);
});
</script>
